requirements.txt
pybit==5.0.0
pandas==2.0.3
numpy==1.24.3
pyyaml==6.0.1
scipy==1.10.1
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
python-dateutil==2.8.2